:recycle:  Refactor: 공통 타입 경로 수정
models폴더 삭제, types 폴더 이용
