+++ src/types/index.ts
@@ -1,8 +1,1 @@
 // @공통 type 정의
-
-export interface UserData {
-  userId: string;
-  userName: string;
-  content: string;
-  date: Date;
-}
